Mostly maintenance and polishing + system level exception catching
The backend now sets the imgui context in modules, so it doesn't
need to be passed in each gui function and require the module
programmer to set the imgui context every time.

Serialization has been refactored and is now done using the
Binary_Archive. This is used to load/save the ecs and user
preferences.

When a module function is called from the backend, it is encapsulated
in SEH __try __except (Win32) so all system level exceptions in
modules are caught an reported in the log rather than crashing the
engine.

A basic 2D particle simulation was implemented in a new module. It
currently only renders points, but the concept is there.