Custom container implementations, project loading, file browser improvements
Created custom implementations for Dynamic_Array (no longer just an alias for std::vector)
and Static_Array all with custom iterators and allocators. Making a custom specialized
allocator is simple - create a class that implements the required template functions and
pass it in the template arguments for the Dynamic_Array.

Added the feature to load different projects (directories) as well as save them in the
respective directory. User preferences are saved to a global directory (.user).

Improved the file browser to be able to select folders.

Fixes, polishing & improvements.